Fix closures: add Closure class
The clue was that there was no first-class Closure value! When an Fexpr/Fn
is eval’d, that’s when we need to close over its variables. Previously,
evalling an Fexpr/Fn did nothing, and the closure happened at the call site,
which was wrong.

Also fix a bug where the global binding was being added to the environment
each time the environment was extended.

This adds tests demonstrating a) that a closed-over variable can retain
state between calls of the closure and b) that each time a function is
instantiated the closure is re-formed.