Fix free variable and let bindings
Fexpr was binding too early. It should bind when the Fexpr is evalled, not
before.

Let was only binding once, rather than each time it was executing, thus
failing to create separate instances of bound variables.

Add a test that let-binds a function then defines it with a recursive call.
This is a sufficient workaround for the lack of letrec.