Hak: Add “use” for JavaScript libraries (incomplete)
So far, this doesn’t work properly, as “env” is passed as an extra first
argument to all property functions.

To allow us to use “require”, switch to using CommonJS.

Because of this, rewrite the Mocha tests in Ava, and downgrade execa, as
recent versions are ESM-only.

Also rewrite the bin/ scripts.