Rewrite eval, _value and _toJs methods as monolithic functions.
I have not yet mastered the TypeScript tricks to do mixins (which is what we
really need here), and I’d rather end up with a reference implementation
that is easy to translate into other languages.

The immediate motivation is wanting eval to “link” free variables when it is
invoked, but not on recursive calls.