Fix implementation of RefLocation and Namespace
Namespace should be a map from string to Ref (not to Val).

RefLocation then simply uses these Refs (rather than constructing its own).

This means that ‘interpret’ does not need to double-interpret RefLocations.

Also, in ‘interpret’, we know that the value we get is a Ref, so simply use
its Val rather than recursively interpreting it.