Add ObjLiteral, and interpret all containers the same way
Add Ark and Ursa tests for object literals.

Add missing treatment of PropertyValue’s free vars in Ursa.

Interpreting any container (Obj, List, or Dict) simply returns it.

Adding ObjLiteral means that all the containers are now implemented with a
pair of classes, one for literals and one for actual containers.

Also, in jsToVal, convert Array to ListLiteral and Map to DictLiteral.