Ark: add and use ability for Val to overwrite itself
Add parent reference to `Val`, and methods `addChild` and `setSelf`; set it
in `Val`s that have a parent. A `Val` is identified by a reference to the
parent, and an index into the parent’s `children` array.

Overwrite `SymRef`s with the resolved `Ref`, thereby avoiding an extra
indirection at run-time.