Flag up undefined symbols at the point where they’re compiled
In Ursa, give source location information.

To do this, replace the ‘link’ function with an ‘externalSyms’ member in
‘Environment’, which is used by ‘SymRef.getIndex’ to resolve symbol
references not found on the compilation stack.

To make the error message nice, tweak the Ursa grammar so that the ‘ident’
rule matches an entire identifier in one argument.

Also, simplify some loops removing boundVars from freeVars, and make error
messages consistently start with a capital letter.