Ursa & Ark: add source information to ArkRuntimeError, and use it
Add the `Val` in which a runtime error occurred to `ArkRuntimeError`.

In Ursa, use this to extract source information from the debug object
injected to each object constructed by the Ursa compiler, and hence add
source location info to the error in the new function `runWithTraceback`.

Add a FIXME to handle conversion in `fromJs`, as that does not have the
`Val` available to construct an `ArkRuntimeError`, so its caller will have
to deal with it. Also, for objects created at runtime (in particular,
containers), we’ll need to trace up from the given object to the nearest
parent with a source location.

Also, in the Ursa compiler, inline `makeFn` in its only caller.