Add debug source location stack to ArkState and use it
This is a stack that corresponds to the tree position (not stack frame, that
will come next). Push/pop in each Val.eval() method.

Use this new functionality to add source location to JS-to-Val conversion
errors: fromJs doesn’t know what Val it’s called from.

Also fix an error in `serialize` that was just introduced where we failed to
serialize properly if a `Val` had a debug table without a `name` element.