Fix to implementation of local variables
Let-bound variables are gathered up into a single Let per Sequence. This was
not properly reflected in their treatment at compile time, so the wrong
stack slots were being referenced.

Fix this: add a slack slot to the Environment’s innermost stack frame for
each variable at the start of a Sequence, but with ‘undefined’ instead of a
symbol name. Then, as the variables are introduced, add their names at the
correct point of the sequence.

In Sequence, we need to augment the stack frame before compiling any of the
sequence expressions.

Also, in Use, use this existing binding (which already contains the
let-bound identifier for the “use”) rather than introducing an inner
environment augmented with the new let-bound variable.

Finally, when looking up an identifier in an Environment’s stack frame,
start from the end of the frame, in case more-recently introduced
identifiers shadow less-recently-introduced ones.