Fix let in REPL (fix #19)
It’s no longer possible to use ‘Exp’ rather than ‘Sequence’ as the initial
production, because of the way that let-bound variables are collected
per-Sequence.

The compiler is really confusing three things here:

1. Ergonomics of the code for the REPL. For the REPL we want all lets to
migrate to the top level of the expression parsed. But we can do this simply
by starting at ‘Exp’ as before.

2. Efficiency of compiled code. (We don’t want to have another ArkLet and
code nesting level for each `let` in the source.)

3. Abstract syntax: here, we really do want to keep the ArkLet corresponding
to the source `let`.

So, we should revert to compiling more literally what the source code says,
which will allow partial compilation (e.g. starting with ‘Exp’ as the
initial production) to work naturally. Then we should optimize the generated
code later, e.g. “pulling” ArkLets up to the top of an ArkSequence. (And
similarly for other optimizations such as eliding sequences of length 1.)
We’ll want to do more optimization as time goes on anyway.