Change Object syntax (fixes #36)
Change the syntax for Object initializers to “key = value”, separated by
semi-colons. This unifies the syntax with ‘let’.

Fix #36 by switching the order of Object and Map in the PrimaryExp grammar
production, so “{}” is recognized as an empty map. (This seems to be a bit
more useful in general than an empty object.)

Because of the new syntax for objects, it’s possible nonetheless to have an
empty object literal: `{;}`.