Fold RuntimeStack into ArkState and make it async/continuation-ready
The interpreter does not need a stack, just the current frame. We only want
to know about outer frames for debug purposes. Hence, in ArkState store a
frame and outer frame.

This means that the only updates to an ArkState happen within a function
context, so async and continuation execution of different function states
concurrently can now work.

The ArkState in which an exception is generated must now be stored in the
ArkRuntimeError. This has the side-effect that when ArkFromJsError is thrown
in NativeObject.get, we can’t construct an ArkRuntimeError, as we don’t know
the current ArkState. Hence, remove the code that did this. We will probably
want to revisit how exceptions coming “from JavaScript” are handled later in
a more comprehensive way.