Make Ohm output patching easier to maintain
Make the patch just add our extra code and types, and use sed to change the
type of “this” arguments in semantic actions from NonterminalNode to
ThisNode.

As a result, grammar changes will not break the patch unless they occur
in the top few or bottom few rules, which the patch covers currently.