Fix one FIXME and remove another
Improve the check for accessing a non-existent property of an Object.
Raise an error when a non-existent property is accessed. This means that no
new properties can be assigned to an Object that didn’t exist in the literal
from which it was created.

Don’t change the class hierarchy to reflect a difference between “indexable”
objects (List, String) and those that have properties, as there is no real
difference, especially now that we don’t have indexing syntax.

Return `ArkUndefined`, not `ArkNull` or `undefined`, for undefined
properties. Not returning `ArkNull` means that we can treat it like any
other object in the properties check.

Remove error for “Uninitialized symbol” in ArkLexp, as in fact there’s
nothing wrong per se with having an uninitialized symbol as an lvalue; we
could be about to assign to it.

Also, simplify Ursa basics tests that raise a compile-time error: don’t
attempt to run the code, as we will never get that far.

Also, don’t print the traceback heading when an error occurs at the top
level and there’s no traceback.