Convert intrinsics to methods
Fix a bug in fromJs(): list values were not converted to ArkVal.

Remove the standalone tests of JSON inputs. These need to be
regenerated or edited too often, and it's tedious to inspect them to
make sure they are "correct". We still compile the Ursa tests to JSON
and re-run them.