Make ‘or’ evaluate to ‘true’ when LHS is truthy
Ursa will be statically typed, so ‘if’ conditions and the arguments to ‘and’
and ‘or’ will have to be booleans in future anyway.

This simplifies flat code generation for ‘ifElseBlock’, as the ‘then’ part
is now always present.

As a result, remove ‘ArkLetCopyInst’ completely: two uses were not needed,
and this change eliminates the last one.