Fix ending generators
Keep the 'continuation' reference in the frame in which the generator
is running, not its caller, as otherwise a) there can only be one
generator active in a frame, and b) it will be terminated on any
return to the frame!

Add a new Rosetta Code example that detected this problem (Hailstone
sequence). This example is not quite in the right form for Rosetta
Code, so I'll update it next.