Separate methods from properties
Use ArkInvoke for all method calls. This required updating the Ark
tests to use 'invoke' rather than 'call'; and in 'fromJs' we must know
when we are binding a function as a method, and drop its first
argument passed from Ursa ('this' is bound via .bind()).

Tweak grammar to require that the open parenthesis of a function or
method call must not come after a newline. This issue was surfaced by
needing to parenthesise calls to properties. Improve the documentation
of the 'sc' rule. Also, move vertical tab and form feed to line
terminators (since they involve vertical movement).

Simplify the implementation of 'jslib'.

Add explicit 'thisVal' argument to NativeFn methods.