Make Locations unique
Only ever construct one Location for each actual location.

Add a `location` field to `ArkNamedLoc` and `ArkBoundVar`, instead of
copying information from a location into them.

This will enable the type-checker to get the correct type of a location when
it’s computed after references to the location have already been compiled.