Add expected types to “basics” test results
The “examples” tests are run via the command line, and therefore we cannot
test the type of the result (unless/until we encode types directly in JSON).
However, these tests already have an expected concrete test result, so
perforce the result has to be the right type (assuming that the conversion
to JSON is sound!).