Add function and loop references to return/yield/break expressions
So we can type-check `return`, `yield`, and `break` expressions, add a
reference to their containing function or loop.

To achieve this, construct function and loop objects before compiling their
bodies. This enables us to keep track of the function or loop we’re inside.

Make ArkExp non-abstract (no other code change) so that we can use one as a
placeholder body for functions and loops while they’re being compiled.