Improve type-checking of functions, loops and ifs
Type-check ‘break’ and ‘return’ expressions, and function bodies (against
the return type).

The type of a ‘loop’ expression is the union of the types of its ‘break’
expressions, if any, otherwise a new type for ‘non-terminating’.

The type of an ‘if’ expression is the union of the types of its then and, if
any, else arm.

Add FIXMEs for checking generators and ‘yield’, which we don’t yet type
properly.